tests/wpt/web-platform-tests/fullscreen/api/document-onfullscreenerror.html

<!DOCTYPE html>
<title>Document#onfullscreenerror</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function(t)
{
    var sync = true;
    assert_equals(document.onfullscreenerror, null, "initial onfullscreenerror");
    document.onfullscreenerror = t.step_func_done(function(event) {
        assert_false(sync);
    });
    var e = document.createElement('span');
    e.requestFullscreen();
    sync = false;
});
</script>
