LayoutTests/imported/w3c/web-platform-tests/fullscreen/api/document-onfullscreenerror-expected.txt


FAIL Document#onfullscreenerror assert_equals: initial onfullscreenerror expected (object) null but got (undefined) undefined
